Minor Improvement
* 'for' loop replaceable with enhanced 'for' loop
* Unnecessary unboxing
* Explicit type can be replaced with <>
* Statement lambda can be replaced with expression lambda
* Call to 'Arrays.asList()' with too few arguments

+++ commons-geometry-core/src/test/java/org/apache/commons/geometry/core/internal/IteratorTransformTest.java
@@ -70,7 +70,7 @@ public class IteratorTransformTest {
         @Override
         protected void acceptInput(final Integer input) {
             // filter out odd integers
-            final int value = input.intValue();
+            final int value = input;
             if (value % 2 == 0) {
                 final char[] chars = (value + "").toCharArray();
